Revert "fixed glitchy player"
This reverts commit 353af5a222e65fa7f04a1929e2df52a1f714d4a3.

+++ Gradded-Unit/src/com/FuturePixels/Drawables/Levels/PlatForm.java
@@ -25,7 +25,7 @@ public class PlatForm extends IDrawable {
         super();
     }
 
-    public PlatForm(Vector position, double RadianRotation) {
+    public PlatForm(Vector position, float RadianRotation) {
         super();
         setPosition(position);
         setRotation(RadianRotation);
